Move and rename DispenserBehaviourUtil
